requirements.txt
Flask==3.0.0
Flask-Dance==7.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
PyPDF2==3.0.1
fpdf==1.7.2
Werkzeug==3.0.1
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
Pillow==10.2.0
gunicorn==21.2.0